refactor(extension-common): 広告リンク検証の Service Worker 配線を link-verification へ移す
setupBackground は 320 行で、サイドパネル / バッジ / インストール時初期化 /
広告リンク検証 / フレーム CAS / Basic Auth の 6 話題のリスナーを並べていた。
うち広告リンク検証が 130 行と約 4 割を占め、この話題だけは link-verification/
に 7 ファイルの自前ディレクトリを持ちながら、Service Worker 側の配線が他 5 話題
と同居していた。

link-verification/background.ts に setupLinkVerification(buildWarningUrl) と
して切り出し、setupBackground から呼ぶ。content-script 側を
setupAdClickDetection として切り出したのと対称になる。setupBackground は 241 行
になった。

なお、これは 1de4d658 でおこなった chrome.tabs.onRemoved の統合を部分的に戻す。
あのときはレビューの「同じイベントに独立した 2 リスナーは実行順序が保証され
ない」という指摘に従ったが、2 つはバッジのタイマーと検証状態という互いに素な
対象を破棄するだけで、順序が問題になる失敗を確かめていなかった。それぞれの
状態を持つモジュールが自分のクリーンアップを登録する形のほうが筋が通る。

tab-badge の分割は今回のスコープ外とした。
